+++ srcs/ft_putnbr.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yosherau <yosherau@student.42kl.edu.my>    +#+  +:+       +#+        */
+/*   By: yosherau <yosherau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 17:50:51 by yosherau          #+#    #+#             */
-/*   Updated: 2024/11/16 16:43:59 by yosherau         ###   ########.fr       */
+/*   Updated: 2024/11/17 17:40:41 by yosherau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "srcs.h"
 
-int	ft_putnbr(int nbr)
+int	ft_putnbr(long nbr)
 {
 	char	*str;
 	int		length;
